requirements.txt
Django==4.0
psycopg2-binary==2.9.1
Pillow==8.4.0
requests==2.25.1
django-admin-thumbnails==0.2.6
python-decouple==3.6
django-admin-honeypot==1.1.0
django-session-timeout==0.1.0
django-filter==22.1
django-environ==0.9.0
gunicorn==20.1.0
asgiref==3.5.2
django-appconf==1.0.4
django-icons==21.3
django-imagekit==4.1.0
djangorestframework==3.12.2
imgkit==1.2.2
pilkit==2.0
pytz==2022.1
six==1.15.0
sqlparse==0.4.2
Collectfast==2.2.0
python-slugify==5.0.0

django-rest-framework==0.1.0
django-session-security==2.6.1
django-session-timeout==0.1.0
django-summernote
djangorestframework-simplejwt==5.2.0